requirements.txt
aioserial==1.3.0
async-timeout==3.0.1
click==7.1.2
fastapi==0.61.1
h11==0.9.0
hiredis==1.1.0
pydantic==1.6.1
pyserial==3.4
SQLAlchemy==1.3.19
starlette==0.13.6
uvicorn==0.11.8
websockets==8.1
